Class cancelled notification
This app use twilio api to provide a push notification system for
professors to notify all students when the classes are cancelled.
This app is developed as a web app so far using HTML, CSS, CSS3 Mysql,
PHP and a few other technologies. This app will soon be available on
android and apple devices for schools who are subscribers.